6 files modified, 6 added, 1 deleted

+++ schema/page.ts
@@ -1,3 +1,5 @@
+import { Condition } from '../types/condition';
+
 export default {
   $schema: 'http://json-schema.org/draft-07/schema#',
   type: 'object',
@@ -5,22 +7,31 @@ export default {
     heading: {
       title: 'Heading',
       type: 'string',
+      placeholder: 'Enter heading here',
     },
     description: {
       title: 'Description',
       type: 'string',
+      placeholder: 'Enter description here',
     },
     primaryButtonText: {
       title: 'Primary Button Text',
       type: 'string',
+      placeholder: 'Enter primary button text here',
     },
     secondaryButtonText: {
       title: 'Secondary Button Text',
       type: 'string',
+      placeholder: 'Enter secondary button text here',
     },
     truncateDescription: {
       title: 'Truncate Description',
       type: 'boolean',
+      condition: {
+        dependencyField: 'description',
+        operation: 'greater_than',
+        value: 100,
+      },
     },
   },
   required: [
@@ -37,6 +48,8 @@ export type SchemaType = {
     [key: string]: {
       title: string;
       type: string;
+      condition?: Condition;
+      placeholder?: string;
     };
   };
   required: string[];
